+++ src/test/kotlin/com/zopa/ktor/opentracing/util/TestTracer.kt
@@ -1,0 +1,6 @@
+package com.zopa.ktor.opentracing.util
+
+import com.zopa.ktor.opentracing.ThreadContextElementScopeManager
+import io.opentracing.mock.MockTracer
+
+val mockTracer = MockTracer(ThreadContextElementScopeManager())
